+++ src/io/Properties.js
@@ -13,7 +13,6 @@ export default function Properties(activity, propertiesDef, context) {
     dataOutputObjects: [],
   };
 
-
   for (const {id, ...def} of propertiesDef.values) {
     const source = {
       id,
@@ -62,6 +61,9 @@ const proto = Properties.prototype;
 
 proto.activate = function activate(message) {
   if (this[kConsuming]) return;
+  if (message.fields.redelivered && message.fields.routingKey === 'run.start') {
+    this._onActivityEvent('activity.enter', message);
+  }
 
   if (message.fields.redelivered && message.content.properties) {
     this._onActivityEvent('activity.extension.resume', message);